moves test cases to Cart.test.tsx
