+++ hackathon_site/dashboard/frontend/src/api/api.ts
@@ -82,7 +82,7 @@ export const get = <T>(
     return axios.get<T>(`${SERVER_URL}/${uri}`, makeConfig());
 };
 
-export const post = (uri: string, data: any) => {
+export const post = (uri: string, data?: any) => {
     uri = cleanURI(uri);
     return axios.post(`${SERVER_URL}/${uri}`, data, makeConfig());
 };
